Top-k Utility Functions
Score Predictor:
- Given a score_range with histograms representing the remaining score
distributions of all lists, this function calculates the probability
that the worst_score of the score_range will be able to pass the target
score passed in as a parameter.

Feature Correlation:
- Given that an object has been seen in one or more lists, this function
calculates the probability that it will be seen in the list passed in as
a parameter based on pre-computed contingency tables.
- Note that for our IMDB data, this is always 1.

Selectivity Estimator:
- This function gives the probability that the object passed in will be
seen in the lists it has not been seen in yet.
- Note that for our IMDB data, this is always 1.

Miscellaneous:
- Merged the change to make score_range accept an arbitrary number of
lists.

Bug Fixes:
- In the score_range class, discovered_fields_ now properly initializes
to all -1.